moar backup

+++ day05/ex04/ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: elebouch <elebouch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/08/07 18:34:51 by elebouch          #+#    #+#             */
-/*   Updated: 2017/08/12 11:16:58 by elebouch         ###   ########.fr       */
+/*   Updated: 2017/08/15 08:50:37 by elebouch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,14 +22,14 @@ int		ft_strlen(char *str)
 
 char	*ft_strncpy(char *dest, char *src, unsigned int n)
 {
-	int i;
-	char *cpy;
-	int k;
+	int		i;
+	char	*cpy;
+	int		k;
 
 	cpy = dest;
 	i = 0;
 	k = ft_strlen(src);
-	while (i < n)
+	while ((unsigned)i < n)
 	{
 		if (i < k)
 			*(dest++) = *(src++);
